Rename PromptChatMessages to PromptChatMessagesCard
