+++ app/src/pages/prompt/promptVersionLoader.tsx
@@ -21,7 +21,7 @@ export async function promptVersionLoader(args: LoaderFunctionArgs) {
           id
           ... on PromptVersion {
             ...PromptInvocationParameters__main
-            ...PromptChatMessages__main
+            ...PromptChatMessagesCard__main
             ...PromptCodeExportCard__main
             ...PromptModelConfigurationCard__main
             description
